requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.45.1
pyyaml==6.0.1
requests==2.31.0
sanic==23.6.0
aiohttp
python-dotenv==1.0.0